chore: auth resource route relocated
